Add integrations for TypeScript and TSLint
TypeScript can't really run as a linter within arc since I believe linters need
to operate on individual files and are invoked separately for each file.
Instead, I made a unit test runner that ignores the list of files and has "type
checking" as a single test.

For TSLint integration, I was able to find an existing TSLint class on GitHub,
and I modified it to also support autofixing (largely basing it off of the
ESLint autofixing code I wrote in #1).

Some limitations with all of this:
* arc says "running unit tests" and "unit tests passed", etc, when it's just
  doing typechecking. Since we've already forked it, we could just change those
  strings, but I'd rather not diverge from upstream too much.
* arc conveniently has a place to put the running time of the "test" (i.e.
  typechecking), which will make it more obvious if it later gets out of hand.
  Inconveniently, it has a hard-coded rule that a test running for >500ms means
  that it's very slow, so the time always is formatted in red.
* Running each file independently is awkward for tslint. Some rules (like "when
  you use `await`, it must be on a promise") use typechecking information, which
  isn't available when run on an isolated file. So if your lint runner requires
  a separate invocation per file, then you either need to do a codebase-wide
  typecheck per file (slow) or limit checking to rules that don't need type
  information (incomplete). For now, (I'm pretty sure) it's doing the incomplete
  option and CI runs the full set of rules.