Add a unit test engine that runs `bin/dev check`
This is based off of the TypeScriptTestEngine that I wrote earlier, but now just
runs a command of the form `bin/dev check --arc file1.js file2.py`. The actual
`bin/dev check` command will be added to aurelia in a later commit, but this is
needed for a smooth transition.