Add a unit test engine that runs `bin/dev check` (#6)
The plan is to replace most our linting/checking via arc with our own script called
`bin/dev check`, which should allow for better performance and give us more
flexibility. This change just adds a simple arc wrapper that calls through to that
script, and a future aurelia commit will reconfigure aurelia to use that script instead
of the other linting.

More details in this quip doc: https://benchling.quip.com/enjVAikYgEq2/2018-07-03-Making-arc-diff-faster

This is based off of the TypeScriptTestEngine that I wrote earlier, but now just
runs a command of the form `bin/dev check --arc file1.js file2.py`. The actual
`bin/dev check` command will be added to aurelia in a later commit, but this is
needed for a smooth transition.